Fix SELinux Read/Write Mount Point
This was an issue for naftulikay/xenial-vm as well, basically on
SELinux Docker hosts, Docker containers get /sys/fs/selinux mounted
read/write. Since this is largely a farce, everything freaks out on
Ubuntu machines and it won't handle it well.

This fixes that by adding an early stage script to remount the
filesystem in read-only mode at a very early boot stage. This will
allow Ansible to install packages, etc.